secure messaging work
